+++ pkg/providers/mtlstruststore/data_akamai_mtlstruststore_ca_set_associations.go
@@ -3,7 +3,6 @@ package mtlstruststore
 import (
 	"context"
 	"fmt"
-	"regexp"
 
 	"github.com/akamai/AkamaiOPEN-edgegrid-golang/v12/pkg/mtlstruststore"
 	"github.com/akamai/terraform-provider-akamai/v9/pkg/meta"
@@ -102,7 +101,8 @@ func (d *caSetAssociationsDataSource) Schema(_ context.Context, _ datasource.Sch
 				Computed:    true,
 				Validators: []validator.String{
 					stringvalidator.ExactlyOneOf(path.MatchRoot("id"), path.MatchRoot("name")),
-					stringvalidator.RegexMatches(regexp.MustCompile(`\S{3,}`), "must not be empty or only whitespace"),
+					stringvalidator.LengthBetween(3, 64),
+					stringvalidator.RegexMatches(mtlstruststore.CASetNameRegex, mtlstruststore.CASetNameDescription),
 				},
 			},
 			"properties": schema.ListNestedAttribute{