+++ pkg/providers/mtlstruststore/testdata/TestDataCASetActivation/ca_set_name.tf
@@ -4,5 +4,5 @@ provider "akamai" {
 
 data "akamai_mtlstruststore_ca_set_activation" "test" {
   id          = 321
-  ca_set_name = "test name"
+  ca_set_name = "test_name"
 }