+++ pkg/providers/mtlstruststore/testdata/TestResCASet/create_empty_description.tf
@@ -4,7 +4,9 @@ provider "akamai" {
 
 resource "akamai_mtlstruststore_ca_set" "test" {
   name                = "set-1"
-  allow_insecure_sha1 = true
+  description         = ""
+  allow_insecure_sha1 = false
+  version_description = "Initial version for testing"
 
   certificates = [
     {