+++ pkg/providers/mtlstruststore/testdata/TestResCASet/create_no_all_descriptions.tf
@@ -4,7 +4,7 @@ provider "akamai" {
 
 resource "akamai_mtlstruststore_ca_set" "test" {
   name                = "set-1"
-  allow_insecure_sha1 = true
+  allow_insecure_sha1 = false
 
   certificates = [
     {